% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auc_curve_with_rf_model_modular.R
\name{get_auc_curve_with_rf_model_modular}
\alias{get_auc_curve_with_rf_model_modular}
\title{Compute and Plot AUC Curve with Random Forest Model}
\usage{
get_auc_curve_with_rf_model_modular(ml_formatted_scores_df, best.m)
}
\arguments{
\item{ml_formatted_scores_df}{A data frame containing the training data. If \code{NULL}, data will be fetched from the database.}

\item{best.m}{The 'mtry' hyperparameter for Random Forest. If \code{NULL}, it is determined by the function.}
}
\value{
This function does not return any explicit value. It generates:
\itemize{
\item The AUC (Area Under the Curve) printed to the console.
\item A ROC curve plot with the calculated AUC value.
\item Various performance metrics (e.g., True Positive Rate, False Positive Rate), displayed in the plot.
}
}
\description{
This function trains a Random Forest model, computes the ROC curve, and
calculates the AUC (Area Under the Curve). It allows various preprocessing
options, such as imputation, rounding, undersampling, and hyperparameter tuning.
}
\details{
The function requires a data frame that is generated from function f8. The model is trained using the Random Forest algorithm, and
performance is evaluated via the ROC curve and AUC metric. After training the model,
predictions are made, and the AUC is calculated and visualized with a ROC curve plot.
}
\examples{
\dontrun{
# Example 1: Using real data from the database
#This is a placeholder example. Replace the path with a valid database.
 get_auc_curve_with_rf_model(Data = NULL, path_db = "path/to/database.db", rat_studies = TRUE, reps = 10,
                             holdback = 0.75, error_correction_method = "Prune")

# Example 2: Using sample data (if applicable)
 get_auc_curve_with_rf_model(Data = sample_data, path_db = NULL, rat_studies = FALSE, reps = 5,
                             holdback = 0.5, error_correction_method = "CorrectMethod")
}
}
\seealso{
\code{randomForest}, \code{ROCR}
}
