% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_parameter_rf_model.R
\name{get_fixed_parameter_rf_model}
\alias{get_fixed_parameter_rf_model}
\title{get_random_forest_model}
\usage{
get_fixed_parameter_rf_model(
  Liver_get_liver_om_lb_mi_tox_score_list,
  not_Liver_get_liver_om_lb_mi_tox_score_list
)
}
\arguments{
\item{Liver_get_liver_om_lb_mi_tox_score_list}{Mandatory, character \cr
Studyid number}

\item{not_Liver_get_liver_om_lb_mi_tox_score_list}{Mandatory, character \cr
path of database}
}
\value{
score
}
\description{
get_random_forest_model
}
\examples{
\dontrun{
get_liver_lb_score(studyid='1234123', database_path = dbtoken)
}
}
