% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copy__Ofcolumn_harmization.R,
%   R/column_harmization.R
\name{get_harmonized_column}
\alias{get_harmonized_column}
\title{get_random_forest_model}
\usage{
get_harmonized_column(
  data_frame = NULL,
  Liver_get_liver_om_lb_mi_tox_score_list,
  not_Liver_get_liver_om_lb_mi_tox_score_list
)

get_harmonized_column(
  data_frame = NULL,
  Liver_get_liver_om_lb_mi_tox_score_list,
  not_Liver_get_liver_om_lb_mi_tox_score_list
)
}
\arguments{
\item{data_frame}{Mandatory, character \cr
Studyid number}

\item{Liver_get_liver_om_lb_mi_tox_score_list}{Mandatory, character \cr
Studyid number}

\item{not_Liver_get_liver_om_lb_mi_tox_score_list}{Mandatory, character \cr
path of database}
}
\value{
score

score
}
\description{
get_random_forest_model

get_random_forest_model
}
\examples{
\dontrun{
get_liver_lb_score(studyid='1234123', database_path = dbtoken)
}
\dontrun{
get_liver_lb_score(studyid='1234123', database_path = dbtoken)
}
}
