% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lb_score.R
\name{get_lb_score}
\alias{get_lb_score}
\title{Get LB Score for a Given Study ID}
\usage{
get_lb_score(
  studyid = NULL,
  path_db,
  fake_study = FALSE,
  use_xpt_file = FALSE,
  master_compiledata = NULL,
  return_individual_scores = FALSE,
  return_zscore_by_USUBJID = FALSE
)
}
\arguments{
\item{studyid}{Mandatory, character
The study ID number for which the LB score is calculated.}

\item{path_db}{Mandatory, character
The path to the database containing the necessary data for the calculation.}

\item{fake_study}{Optional, boolean
Indicates whether the study is generated by the SENDsanitizer package. Defaults to \code{FALSE}.}

\item{use_xpt_file}{Mandatory, character
Specifies the path to the XPT (SAS transport) file if it is being used for the study.}

\item{master_compiledata}{Mandatory, character
The path to the compiled master dataset that will be used to calculate the LB score.}

\item{return_individual_scores}{Optional, boolean
If \code{TRUE}, the function will return individual scores for each subject. Defaults to \code{FALSE}.}

\item{return_zscore_by_USUBJID}{Optional, boolean
If \code{TRUE}, the function will return Z-scores by \code{USUBJID}. Defaults to \code{FALSE}.}
}
\value{
numeric
The calculated LB score based on the provided data and parameters.
}
\description{
This function computes the LB score for a given study ID using data stored in a specified database.
It offers various optional parameters to customize the output, such as whether to return individual scores or Z-scores by \code{USUBJID}.
}
\examples{
\dontrun{
# Example usage of the function
get_lb_score(studyid='1234123', path_db='path/to/database.db')
}

}
