% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_liver_livertobw_score.R
\name{get_liver_livertobw_score}
\alias{get_liver_livertobw_score}
\title{get_liver_livertobw_score}
\usage{
get_liver_livertobw_score(
  studyid,
  path_db,
  fake_study = FALSE,
  use_xpt_file = FALSE,
  master_compiledata = NULL,
  bwzscore_BW = NULL,
  return_individual_scores = FALSE
)
}
\arguments{
\item{studyid}{Mandatory, character \cr
Studyid number}

\item{path_db}{Mandatory, character \cr
path of database}

\item{fake_study}{optional, Boolean \cr
whether study generated by SENDsanitizer package}

\item{use_xpt_file}{optional, Boolean \cr
whether use_xpt_file is used on not}

\item{master_compiledata}{optional, Boolean \cr
whether use_xpt_file is used on not}

\item{bwzscore_BW}{optional, Boolean \cr
whether use_xpt_file is used on not}

\item{return_individual_scores}{optional, Boolean \cr
whether use_xpt_file is used on not}
}
\value{
dataframe
}
\description{
get_liver_livertobw_score
}
\examples{
\dontrun{
get_compile_data(studyid='1234123', path_db='path/to/database.db')
}
}
