% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_liver_om_lb_mi_tox_score_list.R
\name{get_liver_om_lb_mi_tox_score_list}
\alias{get_liver_om_lb_mi_tox_score_list}
\title{get_liver_om_lb_mi_tox_score_list}
\usage{
get_liver_om_lb_mi_tox_score_list(
  selected_studies,
  path_db,
  fake_study = FALSE,
  use_xpt_file = FALSE,
  multiple_xpt_folder = FALSE,
  output_individual_scores = FALSE
)
}
\arguments{
\item{selected_studies}{Mandatory, character \cr
Studyid number}

\item{path_db}{Mandatory, character \cr
path of database}

\item{fake_study}{optional, Boolean \cr
whether study generated by SENDsanitizer package}

\item{use_xpt_file}{Mandatory, character \cr
path of database}

\item{multiple_xpt_folder}{Mandatory, character \cr
path of database}

\item{output_individual_scores}{optional, Boolean \cr
whether study generated by SENDsanitizer package}
}
\value{
dataframe
}
\description{
get_liver_om_lb_mi_tox_score_list
}
\examples{
\dontrun{
get_compile_data(studyid='1234123', path_db='path/to/database.db')
}
}
