% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mi_score.R
\name{get_mi_score}
\alias{get_mi_score}
\title{get MI score for a given studyid}
\usage{
get_mi_score(
  studyid = NULL,
  path_db,
  fake_study = FALSE,
  use_xpt_file = FALSE,
  master_compiledata = NULL,
  return_individual_scores = FALSE,
  return_zscore_by_USUBJID = FALSE
)
}
\arguments{
\item{studyid}{Mandatory, character \cr
Studyid number}

\item{path_db}{Mandatory, character \cr
path of database}

\item{fake_study}{optional, Boolean \cr
whether study generated by SENDsanitizer package}

\item{use_xpt_file}{Mandatory, character \cr
Studyid number}

\item{master_compiledata}{Mandatory, character \cr
path of database}

\item{return_individual_scores}{optional, Boolean \cr
whether study generated by SENDsanitizer package}

\item{return_zscore_by_USUBJID}{optional, Boolean \cr
whether study generated by SENDsanitizer package}
}
\value{
score
}
\description{
get MI score for a given studyid
}
\examples{
\dontrun{
get_mi_score(studyid='1234123', path_db='path/to/database.db')
}
}
