% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_forest_model.R
\name{get_random_forest_model}
\alias{get_random_forest_model}
\title{get LB score for a given studyid}
\usage{
get_random_forest_model(
  Liver_get_liver_om_lb_mi_tox_score_list,
  not_Liver_get_liver_om_lb_mi_tox_score_lis
)
}
\arguments{
\item{studyid}{Mandatory, character \cr
Studyid number}

\item{database_path}{Mandatory, character \cr
path of database}
}
\value{
score
}
\description{
get LB score for a given studyid
}
\examples{
\dontrun{
get_liver_lb_score(studyid='1234123', database_path = dbtoken)
}
}
