% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repeat_dose_parallel_studyids.R
\name{get_repeat_dose_parallel_studyids}
\alias{get_repeat_dose_parallel_studyids}
\title{Get Repeat Dose Parallel Study IDs}
\usage{
get_repeat_dose_parallel_studyids(path_db, rat_studies = FALSE)
}
\arguments{
\item{path_db}{A character string representing the file path to the SQLite database. This is a required parameter.}

\item{rat_studies}{A logical flag indicating whether to filter the studies for rats only. Defaults to \code{FALSE}.}
}
\value{
A Data frame having the STUDYID that meet the specified criteria. This includes:
\itemize{
\item STUDYIDs that match both the parallel design and repeat-dose toxicity criteria.
\item Optionally, STUDYIDs that match rat species if \code{rat_studies = TRUE}.
}
}
\description{
This function retrieves study IDs from a database that correspond to parallel-design studies involving repeat-dose toxicity.
It optionally filters the studies for rat species.
}
\examples{
\dontrun{
  # Example without filtering for rat studies
  parallel_repeat_dose_df <- get_repeat_dose_parallel_studyids(path_db = "path/to/database.sqlite")

  # Example with filtering for rat studies
  rat_parallel_repeat_dose_df <- get_repeat_dose_parallel_studyids(path_db = "path/to/database.sqlite", rat_studies = TRUE)
}

}
