% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_treatment_group_&_dose.R
\name{get_treatment_group_amin}
\alias{get_treatment_group_amin}
\title{Get SETCD of treatment group  for a given STUDYID from SQLite Database or \code{.xpt} Files}
\usage{
get_treatment_group_amin(
  studyid = NULL,
  path_db,
  fake_study = FALSE,
  use_xpt_file = FALSE
)
}
\arguments{
\item{studyid}{Character. STUDYID number. Defaults to \code{NULL}.
Required for SQLite databases (\code{use_xpt_file = FALSE}).
Must be \code{NULL} for \code{.xpt} files (\code{use_xpt_file = TRUE}).}

\item{path_db}{Character. Path to the SQLite database file or a folder containing \code{.xpt} files. Mandatory.}

\item{fake_study}{Logical. Whether the study data is generated by the \code{SENDsanitizer} package. Defaults to \code{FALSE}.}

\item{use_xpt_file}{Logical. Whether to retrieve study data from \code{.xpt} files instead of the SQLite database. Defaults to \code{FALSE}.}
}
\value{
a data frame
}
\description{
This function retrieve the treatment_group (SETCD) a given STUDYID using data stored in a specified database.
}
\examples{

\dontrun{
# Example usage of get_mi_score
get_mi_score(studyid = '1234123', path_db = 'path/to/database.db')
}

}
