% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rf_predictions.R,
%   R/predicted_random_forest_model.R
\name{predicted_random_forest_model}
\alias{predicted_random_forest_model}
\title{predicted_random_forest_model}
\usage{
predicted_random_forest_model(dbPath_liver, dbPath_not_liver)

predicted_random_forest_model(dbPath_liver, dbPath_not_liver)
}
\arguments{
\item{dbPath_liver}{Mandatory, character \cr
Studyid number}

\item{dbPath_not_liver}{Mandatory, character \cr
path of database}
}
\value{
random forst model

random forst model
}
\description{
predicted_random_forest_model

predicted_random_forest_model
}
\examples{
\dontrun{
get_mi_score(studyid='1234123', path_db='path/to/database.db')
}
\dontrun{
get_mi_score(studyid='1234123', path_db='path/to/database.db')
}
}
