% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReprTree.R
\docType{package}
\name{reprtree-package}
\alias{SENDQSAR}
\alias{SENDQSAR-package}
\alias{reprtree-package}
\title{Representative trees from ensembles}
\description{
This package implements the concept of representative trees from ensemble
tree-based learners introduced by Banerjee, et al, (2012). Representative trees
are, in some sense, trees in the ensemble which are on average the "closest" to
all the other trees in the ensemble. Several trees can be representative of the
ensemble in general. This package currently implements the d2 metric of tree closeness
(close in prediction) from Banerjee, et al.
}
\examples{
library(reprtree)
rforest <- randomForest(Species~., data=iris)
reptree <- ReprTree(rforest, iris, metric='d2')
plot(reptree, index=1)
}
\references{
M. Banerjee, Y. Ding and A-M Noone (2012) "Identifying representative
trees from ensembles". Statistics in Medicine, 31(15):1601-1616.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aminuldu07.github.io/SENDQSAR/}
  \item \url{https://github.com/aminuldu07/SENDQSAR}
  \item Report bugs at \url{https://github.com/aminuldu07/SENDQSAR/issues}
}

}
\author{
\strong{Maintainer}: Md Aminul Islam Prodhan \email{aminuldu07@gmail.com} (\href{https://orcid.org/0000-0003-0205-1250}{ORCID})

}
