% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bw_score.R
\name{get_bw_score}
\alias{get_bw_score}
\title{Calculate BW Score for a Given Study ID}
\usage{
get_bw_score(
  studyid = NULL,
  path_db,
  fake_study = FALSE,
  use_xpt_file = FALSE,
  master_compiledata = NULL,
  return_individual_scores = FALSE,
  return_zscore_by_USUBJID = FALSE
)
}
\arguments{
\item{studyid}{Mandatory, character \cr
The study ID for which the BW score is to be calculated. If \code{NULL}, all studies in the database are analyzed.}

\item{path_db}{Mandatory, character \cr
The path to the database file containing the study data.}

\item{fake_study}{Optional, Boolean \cr
Indicates whether the study was generated by the \code{SENDsanitizer} package. Default is \code{FALSE}.}

\item{use_xpt_file}{Mandatory, Boolean \cr
If \code{TRUE}, the function uses \code{.xpt} files for processing the study data. Default is \code{FALSE}.}

\item{master_compiledata}{Optional, character \cr
The path to an additional database or compiled data file for analysis. If \code{NULL}, only the primary database is used.}

\item{return_individual_scores}{Optional, Boolean \cr
If \code{TRUE}, the function returns individual scores for each record in the study. Default is \code{FALSE}.}

\item{return_zscore_by_USUBJID}{Optional, Boolean \cr
If \code{TRUE}, the function returns z-scores calculated by \code{USUBJID} (unique subject identifiers). Default is \code{FALSE}.}
}
\value{
A \code{data.frame} containing the calculated BW scores. The structure of the output depends on the provided parameters:
\itemize{
\item If \code{return_individual_scores = TRUE}: Returns individual scores for each record.
\item If \code{return_zscore_by_USUBJID = TRUE}: Returns z-scores by \code{USUBJID}.
\item Otherwise, a summarized BW score for the specified \code{studyid}.
}
}
\description{
The \code{get_bw_score} function calculates the Bayesian Weighted (BW) score for a specified study ID using data from a provided database.
It supports optional parameters for fine-tuning the analysis and offers the flexibility to return individual scores or z-scores by \code{USUBJID}.
}
\examples{
\dontrun{
# Example 1: Basic usage
get_bw_score(studyid = '1234123', path_db = 'path/to/database.db')

# Example 2: Include individual scores
get_bw_score(studyid = '1234123', path_db = 'path/to/database.db', return_individual_scores = TRUE)

# Example 3: Include z-scores by USUBJID
get_bw_score(studyid = '1234123', path_db = 'path/to/database.db', return_zscore_by_USUBJID = TRUE)
}

}
