% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compile_data.R
\name{get_compile_data}
\alias{get_compile_data}
\title{Retrieve and Compile Study Data from SQLite Database or \code{.xpt} Files}
\usage{
get_compile_data(
  studyid = NULL,
  path_db,
  fake_study = FALSE,
  use_xpt_file = FALSE
)
}
\arguments{
\item{studyid}{Character. STUDYID number. Defaults to \code{NULL}.
Required for SQLite databases (\code{use_xpt_file = FALSE}).
Must be \code{NULL} for \code{.xpt} files (\code{use_xpt_file = TRUE}).}

\item{path_db}{Character. Path to the SQLite database file or a folder containing \code{.xpt} files. Mandatory.}

\item{fake_study}{Logical. Whether the study data is generated by the \code{SENDsanitizer} package. Defaults to \code{FALSE}.}

\item{use_xpt_file}{Logical. Whether to retrieve study data from \code{.xpt} files instead of the SQLite database. Defaults to \code{FALSE}.}
}
\value{
A data frame containing the following columns:
\itemize{
\item \code{STUDYID} - Study identifier.
\item \code{USUBJID} - Unique subject identifier.
\item \code{Species} - Species of the test subject.
\item \code{SEX} - Sex of the test subject.
\item \code{ARMCD} - Code representing specific Trial Arm (e.g., Low Dose, Mid Dose).
\item \code{SETCD} - Code representing distinct "groups" or "trial sets".
}
}
\description{
Retrieves, cleans, and compiles data for a specified STUDYID
from either a SQLite database or \code{.xpt} files in a given folder.
}
\examples{
\dontrun{
# Retrieve data for a specific study ID from the database
get_compile_data(studyid = '1234123',
                 path_db = 'path/to/database.db',
                 fake_study = FALSE,
                 use_xpt_file = FALSE)

# Retrieve data from all XPT files in a folder
get_compile_data(studyid = NULL,
                 path_db = 'path/to/folder',
                 fake_study = FALSE,
                 use_xpt_file = TRUE)
}
}
