% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_livertobw_score.R
\name{get_livertobw_score}
\alias{get_livertobw_score}
\title{Calculate Liver-to-Body-Weight ratio Z-Scores for each animal in a given STUDYID from SQLite Database or \code{.xpt} Files}
\usage{
get_livertobw_score(
  studyid = NULL,
  path_db,
  fake_study = FALSE,
  use_xpt_file = FALSE,
  master_compiledata = NULL,
  bwzscore_BW = NULL,
  return_individual_scores = FALSE,
  return_zscore_by_USUBJID = FALSE
)
}
\arguments{
\item{path_db}{Character. Path to the SQLite database file or a folder containing \code{.xpt} files. Mandatory.}

\item{fake_study}{Logical. Whether the study data is generated by the \code{SENDsanitizer} package. Defaults to \code{FALSE}.}

\item{use_xpt_file}{Logical. Whether to retrieve study data from \code{.xpt} files instead of the SQLite database. Defaults to \code{FALSE}.}

\item{master_compiledata}{Optional, character \cr
If \code{master_compiledata} is not supplied (i.e., \code{NULL}), the function will automatically call the \code{get_compile_data} function to calculate it.}

\item{bwzscore_BW}{Optional, data.frame. \cr
if \code{bwzscore_BW} is not supplied (i.e., \code{NULL}), the function will automatically call the \code{get_bw_score} function to calculate it.}

\item{return_individual_scores}{Optional, logical \cr
If TRUE, the function returns individual scores for each domain by averaging the scores of all subjects/animals (\code{USUBJID}) in the study. Default is \code{FALSE}.}

\item{return_zscore_by_USUBJID}{Optional, logical \cr
If \code{TRUE}, the function returns Z-scores for each animal/subject by \code{USUBJID}. Default is \code{FALSE}.}
}
\value{
A data frame containing liver-to-body-weight ratio z-scores:
\itemize{
\item If \code{return_individual_scores = TRUE}: Returns averaged Z-scores for each  domain per \code{studyid}.
\item If \code{return_zscore_by_USUBJID = TRUE}: Returns Z-score for each animal/subject by \code{USUBJID} for each domain per \code{studyid}.
\item Otherwise, a summarized BW score for the specified \code{studyid}.
}
}
\description{
This function computes liver-to-body-weight ratio  (livertoBW) of each animal and their corresponding z-scores from study data.
It supports retrieving data from SQLite databases or \code{.xpt} files and provides flexible options for output formats.

#' @param studyid Character. STUDYID number. Defaults to \code{NULL}.
Required for SQLite databases (\code{use_xpt_file = FALSE}).
Must be \code{NULL} for \code{.xpt} files (\code{use_xpt_file = TRUE}).
}
\examples{
\dontrun{
# Example 1: Default averaged scores
result <- get_livertobw_score(
  studyid = '1234123',
  path_db = 'path/to/database.db'
)
head(result)

# Example 2: Individual scores by study
result <- get_livertobw_score(
  studyid = '1234123',
  path_db = 'path/to/database.db',
  return_individual_scores = TRUE
)
head(result)

# Example 3: Z-scores by USUBJID
result <- get_livertobw_score(
  studyid = '1234123',
  path_db = 'path/to/database.db',
  return_zscore_by_USUBJID = TRUE
)
head(result)
}

}
