% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mi_score.R
\name{get_mi_score}
\alias{get_mi_score}
\title{Get MI score for a given studyid}
\usage{
get_mi_score(
  studyid = NULL,
  path_db,
  fake_study = FALSE,
  use_xpt_file = FALSE,
  master_compiledata = NULL,
  return_individual_scores = FALSE,
  return_zscore_by_USUBJID = FALSE
)
}
\arguments{
\item{studyid}{Mandatory, character \cr
The study ID number for the clinical study.}

\item{path_db}{Mandatory, character \cr
The file path to the database that contains the study data.}

\item{fake_study}{Optional, logical \cr
If TRUE, the function assumes that the study data was generated by the SENDsanitizer package. Default is FALSE.}

\item{use_xpt_file}{Mandatory, logical \cr
If TRUE, indicates that an XPT file should be used instead of a database for analysis.}

\item{master_compiledata}{Mandatory, character \cr
The path to the master compile data, often used to supplement or compile data from multiple sources.}

\item{return_individual_scores}{Optional, logical \cr
If TRUE, the function returns individual MI scores for each participant. Default is FALSE.}

\item{return_zscore_by_USUBJID}{Optional, logical \cr
If TRUE, the function returns the Z-scores by \code{USUBJID} (subject identifier). Default is FALSE.}
}
\value{
A numeric vector or data frame containing the MI scores. The format depends on the specified parameters, such as individual scores or aggregated scores.
}
\description{
This function calculates the MI score for a given study using the provided study ID and database. It allows flexibility in terms of returning individual scores, Z-scores, and more. The function is compatible with both SENDsanitizer-generated datasets and standard clinical study databases.
}
\examples{
\dontrun{
# Example usage of get_mi_score
get_mi_score(studyid = '1234123', path_db = 'path/to/database.db')
}

}
